requirements.txt
Flask==1.0.2
RPi.GPIO==0.7.0
urllib3==1.24.1
